cnn-python/requirements.txt
easyocr==1.7.1
opencv-python==4.8.1.78
matplotlib==3.8.2
numpy==1.26.3
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2